move things around for better compatibility
