--- lib/native.d.cts
@@ -1,1 +1,0 @@
-export * from './binding/proj.d.ts';
