--- lib/wasm.d.mts
@@ -1,2 +1,0 @@
-import PROJ from './index.mjs';
-export default PROJ;
