+++ native/index.cts
@@ -1,0 +1,1 @@
+export * from '../swig/proj.d.ts';
