+++ native/index.mjs
@@ -3,7 +3,7 @@
  * import qPROJ from 'proj.js';
  * const PROJ = await qPROJ;
  */
-import _PROJ from './native.cjs';
+import _PROJ from '../swig/proj_wrap.cjs';
 
 const PROJ = Promise.resolve(_PROJ);
