+++ test/shared/quickstart.ts
@@ -1,9 +1,13 @@
 import { assert } from 'chai';
-import type bindings from 'proj.js';
+import qPROJ from 'proj.js';
+
+// This allows for easier access to the TypeScript types
+// which are hidden behind a Promise
+import type { Proj } from 'proj.js';
 
 // These tests are shared between Node.js and the browser
-export default function (_PROJ: typeof bindings) {
-  let PROJ: Awaited<typeof bindings>;
+export default function (_PROJ: typeof qPROJ) {
+  let PROJ: Awaited<typeof _PROJ>;
   before('ensure module has finished loading', async () => {
     PROJ = await _PROJ;
   });
@@ -24,7 +28,7 @@ export default function (_PROJ: typeof bindings) {
     const sourceCRS = authFactoryEPSG.createCoordinateReferenceSystem('4326');
     console.timeEnd('AuthorityFactory.create()');
     console.time('createFromUserInput()');
-    const targetCRS = PROJ.createFromUserInput('+proj=utm +zone=31 +datum=WGS84 +type=crs', dbContext);
+    const targetCRS = PROJ.createFromUserInput('+proj=utm +zone=31 +datum=WGS84 +type=crs', dbContext) as Proj.CRS;
     console.timeEnd('createFromUserInput()');
     console.time('CoordinateOperationFactory.create().createOperations()');
     const list = PROJ.CoordinateOperationFactory.create().createOperations(sourceCRS, targetCRS, coord_op_ctxt);
